Added delete add user review by rewitew id

+++ BackEnd/Echooling/src/Core/Echooling.Aplication/DTOs/CourseReviewDTOs/CreateCourseReviewDto.cs
@@ -11,7 +11,7 @@ namespace Echooling.Aplication.DTOs.CourseReviewDTOs
     {
         public Guid UserId { get; set; }
         public string? Comment { get; set; }
-        public int rate { get; set; }
+        public decimal rate { get; set; }
         public string? Fullname { get; set; }
         public Guid CourseId { get; set; }
     }
